Reworked some header components
