+++ components/Header.vue
@@ -18,19 +18,7 @@
           </div>
 
           <div id="mobile-components" class="-mr-2 -my-2 flex sm:pr-3 md:hidden">
-            <!-- <div class="flex-1 flex justify-end px-2 lg:ml-6">
-              <div class="max-w-lg w-1/8 lg:max-w-xs">
-                <label for="search" class="sr-only">Search</label>
-                <div class="relative">
-                  <div class="absolute inset-y-0 left-0 pl-2 flex items-center pointer-events-none">
-                    <SearchIcon class="h-5 w-5 text-gray-400" aria-hidden="true" />
-                  </div>
-                  <input id="search" name="search"
-                    class="block w-0 focus:w-full pl-5 focus:pl-10 pr-4 py-3 border border-transparent rounded-md leading-5 bg-white dark:bg-dark-bg text-gray-300 placeholder-gray-400 focus:outline-none focus:bg-white dark:focus:bg-MITRE-light-silver focus:border-white dark:focus:border-dark-bg focus:ring-blue-400 dark:focus:ring-dark-bg focus:text-gray-900 sm:text-sm hover:bg-gray-200 dark:hover:bg-gray-700"
-                    placeholder="Search" type="search" />
-                </div>
-              </div>
-            </div> -->
+            <!-- TODO: Mobile Site Search-->
             <PopoverButton
               class="bg-white dark:bg-dark-bg rounded-md p-2 inline-flex items-center justify-center text-gray-400 hover:text-gray-500 hover:bg-gray-100 focus:outline-none focus:ring-2 focus:ring-inset focus:ring-indigo-500">
               <span class="sr-only">Open menu</span>
@@ -66,40 +54,9 @@
           </PopoverGroup>
           <div class="hidden md:flex items-center justify-end flex-1 pr-2 md:w-0">
             <div class="hidden md:flex items-center justify-end flex-1 md:w-0">
-              <!-- Possible Sign in functionalilty -->
-              <!-- <a href="#" class="whitespace-nowrap text-base font-medium text-gray-500 hover:text-gray-900 dark:hover:text-gray-200"> Sign in </a>
-              <a href="#" class="ml-8 whitespace-nowrap inline-flex items-center justify-center px-4 py-2 border border-transparent rounded-md shadow-sm text-base font-medium text-white bg-blue-600 hover:bg-indigo-700"> Sign up </a> -->
-
-              <!-- Uncomment once search is added -->
-              <!-- <div class="flex-1 flex justify-end px-2 lg:ml-6">
-                <div class="max-w-lg w-1/8 lg:max-w-xs">
-                  <label for="search" class="sr-only">Search</label>
-                  <div class="relative pl-0">
-                    <div class="absolute inset-y-0 left-0 pl-2 flex items-center pointer-events-none">
-                      <SearchIcon class="h-5 w-5 text-gray-400" aria-hidden="true" />
-                    </div>
-                    <input id="search" name="search"
-                      class="block w-0 focus:w-full pl-5 focus:pl-10 pr-4 py-3 border border-transparent rounded-md leading-5 bg-white dark:bg-dark-bg text-gray-300 placeholder-gray-400 focus:outline-none focus:bg-white dark:focus:bg-MITRE-light-silver focus:border-white dark:focus:border-dark-bg focus:ring-blue-400 dark:focus:ring-dark-bg focus:text-gray-900 sm:text-sm hover:bg-gray-200 dark:hover:bg-gray-700"
-                      placeholder="Search" type="search" />
-                  </div>
-                </div>
-              </div> -->
+              <!-- TODO: Site search-->
             </div>
-            <div id="theme-toggle" type="button"
-              class="text-gray-500 dark:text-dark-text hover:bg-gray-200 dark:hover:bg-gray-700 focus:outline-none focus:ring-4 focus:ring-gray-200 dark:focus:ring-gray-700 rounded-lg text-sm p-2.5">
-              <div class="relative flex items-center ">
-                <SunIcon :class="selected == 'Light' ? 'h-6 w-6' : 'hidden'" />
-                <MoonIcon :class="selected == 'Dark' ? 'h-6 w-6' : 'hidden'" />
-                <DesktopComputerIcon :class="selected == 'System' ? 'h-6 w-6' : 'hidden'" />
-                <select name="theme" id="theme"
-                  class="absolute appearance-none pr-3 w-full h-fit opacity-0 dark:bg-slate-600 dark:text-white"
-                  @change="switchSelect($event)">
-                  <option :selected="selected == 'Light'">Light</option>
-                  <option :selected="selected == 'Dark'">Dark</option>
-                  <option :selected="selected == 'System'">System</option>
-                </select>
-              </div>
-            </div>
+            <ThemeSwitcher :selected="selected" />
           </div>
         </div>
       </div>
@@ -125,78 +82,23 @@
                 <div class="my-3">
                   <nav class="grid gap-y-8 py-2">
                     <h1 class="font-bold">Framework</h1>
-                    <NuxtLink v-for="item in framework" :key="item.name" :to="item.href" :class="[
-                      '-m-3 p-3 flex items-center rounded-md hover:bg-gray-50 dark:hover:bg-gray-700',
-                      item.href == route.fullPath
-                        ? 'bg-gray-300 dark:bg-gray-700'
-                        : ''
-                    ]">
-                      <component :is="item.icon" class="flex-shrink-0 h-6 w-6 text-blue-600" aria-hidden="true" />
-                      <span class="ml-3 text-base font-medium text-gray-900 dark:text-MITRE-silver">
-                        {{ item.name }}
-                      </span>
-                    </NuxtLink>
+                    <NavMobileLinks :items="framework" />
 
                     <hr class="drop-shadow" />
 
                     <h1 class="font-bold">Libraries</h1>
-                    <NuxtLink v-for="item in libraries" :key="item.name" :to="item.href" :class="[
-                      '-m-3 p-3 flex items-center rounded-md hover:bg-gray-50 dark:hover:bg-gray-700',
-                      item.href == route.fullPath
-                        ? 'bg-gray-300 dark:bg-gray-700'
-                        : ''
-                    ]">
-                      <component :is="item.icon" class="flex-shrink-0 h-6 w-6 text-blue-600" aria-hidden="true" />
-                      <span class="ml-3 text-base font-medium text-gray-900 dark:text-MITRE-silver">
-                        {{ item.name }}
-                      </span>
-                    </NuxtLink>
+                    <NavMobileLinks :items="libraries" />
 
                     <hr class="drop-shadow" />
 
                     <h1 class="font-bold">Resources</h1>
-                    <NuxtLink v-for="item in navigation" :key="item.name" :to="item.href" :class="[
-                      '-m-3 p-3 flex items-center rounded-md hover:bg-gray-50 dark:hover:bg-gray-700',
-                      item.href == route.fullPath
-                        ? 'bg-gray-300 dark:bg-gray-700'
-                        : ''
-                    ]">
-                      <component :is="item.icon" class="flex-shrink-0 h-6 w-6 text-blue-600" aria-hidden="true" />
-                      <span class="ml-3 text-base font-medium text-gray-900 dark:text-MITRE-silver">
-                        {{ item.name }}
-                      </span>
-                    </NuxtLink>
-                    <NuxtLink v-for="item in resources" :key="item.name" :to="item.href" :class="[
-                      '-m-3 p-3 flex items-center rounded-md hover:bg-gray-50 dark:hover:bg-gray-700',
-                      item.href == route.fullPath
-                        ? 'bg-gray-300 dark:bg-gray-700'
-                        : ''
-                    ]">
-                      <component :is="item.icon" class="flex-shrink-0 h-6 w-6 text-blue-600" aria-hidden="true" />
-                      <span class="ml-3 text-base font-medium text-gray-900 dark:text-MITRE-silver">
-                        {{ item.name }}
-                      </span>
-                    </NuxtLink>
+                    <NavMobileLinks :items="navigation" />
+                    <NavMobileLinks :items="resources" />
                   </nav>
                 </div>
               </div>
               <div class="flex items-center justify-between py-6 px-3">
-                <label class="text-gray-900 text-base font-medium dark:text-MITRE-silver">Switch Theme</label>
-                <div
-                  class="relative flex items-center ring-1 ring-slate-900/10 rounded-lg shadow-sm p-2 text-slate-700 font-semibold dark:bg-slate-600 dark:ring-0 dark:highlight-white/5 dark:text-slate-200">
-                  <SunIcon :class="selected == 'Light' ? 'h-6 w-6' : 'hidden'" />
-                  <MoonIcon :class="selected == 'Dark' ? 'h-6 w-6' : 'hidden'" />
-                  <DesktopComputerIcon :class="selected == 'System' ? 'h-6 w-6' : 'hidden'" />
-                  <div class="ml-2">{{ selected }}</div>
-                  <ChevronDownIcon class="w-5 h-6  text-slate-400" />
-                  <select id="theme"
-                    class="absolute appearance-none inset-0 w-full h-full opacity-0 dark:bg-slate-600 dark:text-white"
-                    @change="switchSelect($event)">
-                    <option :selected="selected == 'Light'">Light</option>
-                    <option :selected="selected == 'Dark'">Dark</option>
-                    <option :selected="selected == 'System'">System</option>
-                  </select>
-                </div>
+                <MobileThemeSwitcher :selected="selected" />
               </div>
             </div>
           </PopoverPanel>
@@ -204,30 +106,6 @@
       </transition>
     </Popover>
     <!-- TODO: Check back later on banner -->
-    <!-- <div id="newsBannerDIV" class="relative bg-blue-600">
-      <div class="max-w-7xl mx-auto py-3 px-3 sm:px-6 lg:px-8">
-        <div class="pr-16 text-center px-16">
-          <p class="font-medium text-white">
-            <span class="md:hidden"> We have new changes! </span>
-            <span class="hidden md:inline">
-              Big news! We're excited to announce new changes!
-            </span>
-            <span class="block ml-2 sm:inline-block">
-              <a href="#" class="text-white font-bold underline">
-                Learn more <span aria-hidden="true">&rarr;</span></a>
-            </span>
-          </p>
-        </div>
-        <div class="absolute inset-y-0 right-0 pt-1 pr-1 flex items-start sm:pt-1 sm:pr-2 sm:items-start">
-          <button type="button" id="dismiss-banner-button"
-            class="flex p-2 rounded-md hover:bg-indigo-500 focus:outline-none focus:ring-2 focus:ring-white"
-            @click="hideBanner()">
-            <span class="sr-only">Dismiss</span>
-            <XIcon class="h-6 w-6 text-white" aria-hidden="true" />
-          </button>
-        </div>
-      </div>
-    </div> -->
   </div>
 </template>
 
@@ -247,17 +125,11 @@ import {
   XIcon,
   CubeIcon,
   ClipboardCheckIcon,
-  SearchIcon,
   DotsVerticalIcon,
-  DesktopComputerIcon,
-  MoonIcon,
-  SunIcon,
   PencilAltIcon,
-  PresentationChartBarIcon,
   UserGroupIcon
 } from '@heroicons/vue/outline';
 import { AcademicCapIcon, BookOpenIcon, ChevronDownIcon, RssIcon } from '@heroicons/vue/solid';
-import NavDropdown from './NavDropdown.vue';
 
 ////  Data  ////
 const selected = ref("")
@@ -405,33 +277,6 @@ const resources = ref([
 
 
 ////  Methods  ////
-const switchSelect = (event) => {
-  if (event.target.value === 'Light') {
-    disableDarkMode()
-    localStorage.setItem('user-theme', 'light');
-    selected.value = "Light"
-    setThemeState(selected.value)
-  }
-  else if (event.target.value == 'Dark') {
-    enableDarkMode()
-    localStorage.setItem('user-theme', 'dark');
-    selected.value = "Dark"
-    setThemeState(selected.value)
-  }
-  else if (event.target.value == 'System') {
-    const userIsInDarkModeOS = window.matchMedia('(prefers-color-scheme: dark)').matches
-    if (userIsInDarkModeOS) {
-      enableDarkMode()
-      localStorage.setItem('user-theme', 'dark');
-      selected.value = "System";
-      setThemeState(selected.value)
-    }
-    else {
-      disableDarkMode()
-      selected.value = "System";
-    }
-  }
-}
 const setThemeState = (mode) => {
   localStorage.setItem('theme', mode)
 }
@@ -445,11 +290,6 @@ const enableDarkMode = () => {
 const disableDarkMode = () => {
   document.documentElement.classList.remove('dark');
 }
-const hideBanner = () => {
-  let bannerDIV = document.getElementById("newsBannerDIV")
-  bannerDIV.style.display = "none"
-}
-
 
 ////  Lifecycle  ////
 onMounted(() => {
