+++ components/SignIn.vue
@@ -1,0 +1,18 @@
+<!-- Currently Not In Use (Remove comment after complete)-->
+<template>
+    <a href="#" class="whitespace-nowrap text-base font-medium text-gray-500 hover:text-gray-900 dark:hover:text-gray-200">
+        Sign in </a>
+    <a href="#"
+        class="ml-8 whitespace-nowrap inline-flex items-center justify-center px-4 py-2 border border-transparent rounded-md shadow-sm text-base font-medium text-white bg-blue-600 hover:bg-indigo-700">
+        Sign up </a>
+</template>
+
+<!-- Currently Not In Use -->
+<script setup>
+////  Props  ////
+
+////  Methods  ////
+
+////  Lifecycle  ////
+
+</script>
