+++ components/SiteSearch.vue
@@ -1,0 +1,26 @@
+<!-- Currently Not In Use And Not Complete (Remove comment after complete)-->
+<template>
+    <div class="flex-1 flex justify-end px-2 lg:ml-6">
+        <div class="max-w-lg w-1/8 lg:max-w-xs">
+            <label for="search" class="sr-only">Search</label>
+            <div class="relative pl-0">
+                <div class="absolute inset-y-0 left-0 pl-2 flex items-center pointer-events-none">
+                    <SearchIcon class="h-5 w-5 text-gray-400" aria-hidden="true" />
+                </div>
+                <input id="search" name="search"
+                    class="block w-0 focus:w-full pl-5 focus:pl-10 pr-4 py-3 border border-transparent rounded-md leading-5 bg-white dark:bg-dark-bg text-gray-300 placeholder-gray-400 focus:outline-none focus:bg-white dark:focus:bg-MITRE-light-silver focus:border-white dark:focus:border-dark-bg focus:ring-blue-400 dark:focus:ring-dark-bg focus:text-gray-900 sm:text-sm hover:bg-gray-200 dark:hover:bg-gray-700"
+                    placeholder="Search" type="search" />
+            </div>
+        </div>
+    </div>
+</template>
+
+<!-- Currently Not In Use And Not Complete-->
+<script setup>
+////  Props  ////
+
+////  Methods  ////
+
+////  Lifecycle  ////
+
+</script>
